chore: move file under components/wrapper related to #10
